Move Type model tests
